| 14445    | Use mkdtemp() instead of hardcoded                          |
   | zmedico  | pym/portage/tests/bin/root directory.                       |
   
   | 14446    | Use explicit imports instead of *.                          |
   | zmedico  |                                                             |
   
   | 14447    | Fix binTestsInit() to use correct values for                |
   | zmedico  | PORTAGE_BIN_PATH and PORTAGE_PYM_PATH.                      |
   
   | 14448    | Use catsplit() instead of a regex to generate               |
   | zmedico  | dbapi._categories. Thanks to Marat Radchenko                |
   |          | <marat@slonopotamus.org> for this patch.                    |
   
   | 14449    | Make _test_pty_eof() use non-blocking IO, required for      |
   | zmedico  | Darwin kernel.                                              |
   
   | 14450    | Fix TypeError in clear_screen() in dispatch-conf which      |
   | arfrever | occurs when Python 3 is used (bug #286682).                 |
   
   | 14451    | Make _test_pty_eof() fork when writing to the slave_fd,     |
   | zmedico  | since otherwise it would block on some platforms such as    |
   |          | Darwin.                                                     |
   
   | 14452    | In _test_pty_eof(), call waitpid on the child process only  |
   | zmedico  | after reading all the data from the pty.                    |
   
   | 14453    | Try to avoid blocking on Darwin in _test_pty_eof() by using |
   | zmedico  | slave_fd directly instead of fdopen.                        |
   
   | 14454    | Make _test_pty_eof() call process.spawn() instead of        |
   | zmedico  | os.fork().                                                  |
   
   | 14455    | Fix breakage in file path -> package lookup code.           |
   | zmedico  |                                                             |

svn path=/main/branches/prefix/; revision=14456

+++ pym/portage/tests/bin/setup_env.py
@@ -3,34 +3,45 @@
 # Distributed under the terms of the GNU General Public License v2
 # $Id$
 
+import tempfile
+
 from portage import os
 from portage import shutil
 from portage.tests import TestCase
 from portage.process import spawn
-from portage.const import PORTAGE_BIN_PATH
 
-bindir = os.path.join(os.path.dirname(os.path.dirname(
+basepath = os.path.join(os.path.dirname(os.path.dirname(
 	os.path.abspath(__file__))),
-	"..", "..", "..", "bin", "ebuild-helpers")
-basedir = os.path.join(os.path.dirname(os.path.dirname(
-	os.path.abspath(__file__))), "bin", "root")
-os.environ["D"] = os.path.join(basedir, "image")
-os.environ["T"] = os.path.join(basedir, "temp")
-os.environ["S"] = os.path.join(basedir, "workdir")
-os.environ["PF"] = "portage-tests-0.09-r1"
-os.environ["PATH"] = bindir + ":" + os.environ["PATH"]
-os.environ["PORTAGE_BIN_PATH"] = PORTAGE_BIN_PATH
+	"..", "..", "..")
+bindir = os.path.join(basepath, "bin")
+pymdir = os.path.join(basepath, "pym")
+basedir = None
+env = None
 
 def binTestsCleanup():
+	global basedir
+	if basedir is None:
+		return
 	if os.access(basedir, os.W_OK):
 		shutil.rmtree(basedir)
+		basedir = None
+
 def binTestsInit():
 	binTestsCleanup()
-	os.mkdir(basedir)
-	os.mkdir(os.environ["D"])
-	os.mkdir(os.environ["T"])
-	os.mkdir(os.environ["S"])
-	os.chdir(os.environ["S"])
+	global basedir, env
+	basedir = tempfile.mkdtemp()
+	env = os.environ.copy()
+	env["D"] = os.path.join(basedir, "image")
+	env["T"] = os.path.join(basedir, "temp")
+	env["S"] = os.path.join(basedir, "workdir")
+	env["PF"] = "portage-tests-0.09-r1"
+	env["PATH"] = bindir + ":" + env["PATH"]
+	env["PORTAGE_BIN_PATH"] = bindir
+	env["PORTAGE_PYM_PATH"] = pymdir
+	os.mkdir(env["D"])
+	os.mkdir(env["T"])
+	os.mkdir(env["S"])
+	os.chdir(env["S"])
 
 class BinTestCase(TestCase):
 	def __init__(self, methodName):
@@ -43,7 +54,7 @@ class BinTestCase(TestCase):
 
 def _exists_in_D(path):
 	# Note: do not use os.path.join() here, we assume D to end in /
-	return os.access(os.environ["D"] + path, os.W_OK)
+	return os.access(env["D"] + path, os.W_OK)
 def exists_in_D(path):
 	if not _exists_in_D(path):
 		raise TestCase.failureException
@@ -54,9 +65,10 @@ def xexists_in_D(path):
 def portage_func(func, args, exit_status=0):
 	# we don't care about the output of the programs,
 	# just their exit value and the state of $D
+	global env
 	f = open('/dev/null', 'wb')
 	fd_pipes = {0:0,1:f.fileno(),2:f.fileno()}
-	spawn(func+" "+args, env=os.environ, fd_pipes=fd_pipes)
+	spawn([func] + args.split(), env=env, fd_pipes=fd_pipes)
 	f.close()
 
 def create_portage_wrapper(bin):
@@ -66,9 +78,10 @@ def create_portage_wrapper(bin):
 		return portage_func(*newargs)
 	return derived_func
 
-for bin in os.listdir(bindir):
+for bin in os.listdir(os.path.join(bindir, "ebuild-helpers")):
 	if bin.startswith("do") or \
 	   bin.startswith("new") or \
 	   bin.startswith("prep") or \
 	   bin in ["ecompress","ecompressdir","fowners","fperms"]:
-		globals()[bin] = create_portage_wrapper(bin)
+		globals()[bin] = create_portage_wrapper(
+			os.path.join(bindir, "ebuild-helpers", bin))
